Created check function
checks the position of the players against the finish line. This was initially part of the main code which was problematic if both players passed the finish line on the same 'for' loop. (In that case, player 1 would automatically win because they rolled first, and their position was checked first. This would give the right result but player 2 would still be allowed a final roll. This is not the desired outcome.