Made code robust against non-ascii or non-integer inputs
Previously, entering a blank or non-integer input for a dice roll would cause an error and break the program as the program was expecting to compare an integer input against an array of integers (in dice_val). Now the program compares the string input against an array of strings (updated dice_val). This prevents users from misuing the program